requirements.txt
flask
flask-sqlalchemy
flask-wtf
flask-bcrypt
flask-login
wtforms
itsdangerous